Add day 2 puzzle
It turns out that sprintf doesn't actually print when called from within
a for loop (or maybe an if statement). I have tried to debug part 2 for a
good hour before I figured out that using cat instead of sprintf solved
the issue. smh.

Also, the fact that arrays are 1-indexed in R is dumb. Like really dumb.