Solve day 3 part 2
Holy moly Batman, this was bad. It seems that there is a weird thing going
on with duckdb that messed up the ordering of the numbers. The first was
my fault for not using generate_subscript. The second was some weird thing
with CTEs. It seems that duckdb re-evaluates them every time you call them,
so the joins in gear_ratios were using different tables even if I was calling
the same one. That's weird.